# packages/web-core/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Duck Engine Web Core API",
    "version": "1.0.0",
    "description": "Core web API for Duck Engine (resource-first materials + file assets; scenes/editor next)",
    "contact": {
      "name": "SpaceDucks Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/admin/users/{userId}": {
      "patch": {
        "tags": [
          "Admin",
          "Users"
        ],
        "summary": "Update a user",
        "description": "Requires SUPER_ADMIN. Cannot change your own role or disable yourself.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/admin/resources": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List resources",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional resource kind filter (e.g. standardMaterial)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of resources",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "count"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResourceSummary"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a resource",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "zip"
                ],
                "properties": {
                  "zip": {
                    "type": "string",
                    "format": "binary",
                    "description": "ZIP containing resource.json and slot files (<slot>.<ext>)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created resource",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Resource"
                    },
                    {
                      "$ref": "#/components/schemas/CreateResourceFromZipResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/admin/resources/{resourceId}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get a resource with versions",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceWithVersions"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Patch resource fields",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchResourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "400": {
            "description": "No changes / invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a resource",
        "description": "Hard-deletes the resource (and cascades versions/bindings). Also deletes any now-unreferenced file blobs from disk.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "resourceId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "webcore_auth"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "SUPER_ADMIN",
          "ADMIN",
          "USER"
        ]
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "isActive",
          "sessionVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "isActive": {
            "type": "boolean"
          },
          "sessionVersion": {
            "type": "integer"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PatchUserRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "Resource": {
        "type": "object",
        "required": [
          "id",
          "key",
          "displayName",
          "kind",
          "activeVersion",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "activeVersion": {
            "type": "integer"
          },
          "thumbnailFileAssetId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceSummary": {
        "description": "Resource row with Prisma `_count` for versions.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Resource"
          },
          {
            "type": "object",
            "required": [
              "_count"
            ],
            "properties": {
              "_count": {
                "type": "object",
                "required": [
                  "versions"
                ],
                "properties": {
                  "versions": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        ]
      },
      "ResourceVersion": {
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "version",
          "componentType",
          "componentData",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "componentType": {
            "type": "string"
          },
          "componentData": {
            "type": "string",
            "description": "JSON string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FileAsset": {
        "type": "object",
        "required": [
          "id",
          "fileName",
          "contentType",
          "size",
          "sha256",
          "storagePath",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "storagePath": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceBinding": {
        "type": "object",
        "required": [
          "id",
          "resourceVersionId",
          "slot",
          "fileAssetId",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceVersionId": {
            "type": "string"
          },
          "slot": {
            "type": "string"
          },
          "fileAssetId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResourceBindingWithFile": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResourceBinding"
          },
          {
            "type": "object",
            "required": [
              "fileAsset"
            ],
            "properties": {
              "fileAsset": {
                "$ref": "#/components/schemas/FileAsset"
              }
            }
          }
        ]
      },
      "ResourceVersionWithBindings": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResourceVersion"
          },
          {
            "type": "object",
            "required": [
              "bindings"
            ],
            "properties": {
              "bindings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceBindingWithFile"
                }
              }
            }
          }
        ]
      },
      "ResourceWithVersions": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Resource"
          },
          {
            "type": "object",
            "required": [
              "versions"
            ],
            "properties": {
              "versions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceVersionWithBindings"
                }
              }
            }
          }
        ]
      },
      "CreateResourceRequest": {
        "type": "object",
        "required": [
          "kind",
          "key",
          "displayName"
        ],
        "properties": {
          "kind": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "CreateResourceFromZipResponse": {
        "type": "object",
        "required": [
          "resource",
          "version"
        ],
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/Resource"
          },
          "version": {
            "$ref": "#/components/schemas/ResourceVersion"
          }
        }
      },
      "PatchResourceRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}